Fix GitHub pages and point to new domain.
Github pages was not working due to an error message about cayman not
available as an import. This looks to be due to cayman issue #131. I applied
their suggested workaround and that fixed the error. This pull request
also updates the CNAME file to the new repository and thus fixes #1.